backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
groq==0.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
